Add configure check for "quiet" FP comparisons (isgreater(), etc.)
... and fallback expressions in Macros.h

`isgreater(a, b)` and `(a > b)` actually translate to different
floating point instructions (the instructions differ on how floating
point exception would be generated). While htop doesn't use FP
exception flags for anything, the isgreater() macros and like are
still useful for writing future-proof code that expects NaN in any of
the operands.

Signed-off-by: Kang-Che Sung <explorer09@gmail.com>