Introduce isPercentChart member in MeterClass
This property distinguishes meters that have a (relatively) fixed
"total" value and meters that do not have a definite maximum value.
The former meters would be drawn as a "100% stacked bar" or "graph", and
the latter would have their "total" values updated automatically in bar
meter mode.

This commit is a prerequisite for the "Graph meter dynamic scaling and
coloring" feature.

Signed-off-by: Kang-Che Sung <explorer09@gmail.com>