Obsolete the CUSTOM_METERMODE use in default mode
The use of CUSTOM_METERMODE value in meter default mode was a bad
design. There are no meter that really has a "custom" mode to work with
and currently that value serves as a useless placeholder that hides the
real default mode for a meter.

Replace CUSTOM_METERMODE in `defaultMode` of all meters with the real
intended default modes. Currently only CPU meters and MemorySwapMeter
used this, and their real defaults are BAR_METERMODE.

Remove the special treatment of `defaultMode == CUSTOM_METERMODE` in
Meter_setMode(). Meter_setMode() still calls the `updateMode` function
for custom treatment when it's present for a meter class.

Signed-off-by: Kang-Che Sung <explorer09@gmail.com>