Graph meter dynamic scaling and coloring
Rewrite the entire graph meter drawing code to support dynamic scaling
and colors.

Dynamic scaling is used for meter classes with "isPercentChart" set to
false. The "total" of a graph meter will align to a power of 2 to ease
the management of internal data structures.

The colors of a graph are based on the percentages of item values of
the meter. The rounding differences of each terminal character are
addressed through the different numbers of braille dots.

Due to low resolution of the character terminal, the rasterized colors
may not look nice, but better than nothing. :)

The code is designed with the anticipation that the graph height may be
changeable at runtime. However no UI or option has been implemented for
that yet. The graph height has a limit of 8191 (terminal rows).

Signed-off-by: Kang-Che Sung <explorer09@gmail.com>