Rework Vector_softRemove() and compact() operations
Remove the 'dirty_index' and 'dirty_count' members of Vector object.
Let the caller of the Vector_softRemove() operation track the
'dirtyIndex' and pass it to Vector_compact() when needed.
'dirtyIndex' is supposed to have shorter lifespan than a Vector
instance, and by removing the member we can save object size.

'dirty_count' is redundant and thus removed as well.
A boolean 'isDirty' member is added to a Vector object for ease of
debugging.

Signed-off-by: Kang-Che Sung <explorer09@gmail.com>